- Moved basic image manipulation in filerequester to class KIconFileBase  - Added a cast to remove warning

+++ kiconedit/properties.cpp
@@ -119,7 +119,7 @@ void KIconEditProperties::saveProperties(QWidget *parent)
   config->writeEntry( "Geometry", geom );
 
   debug("Writing bgmode: %d", (int)pprops->backgroundmode);
-  config->writeEntry("BackgroundMode", pprops->backgroundmode);
+  config->writeEntry("BackgroundMode", (int)pprops->backgroundmode);
   config->writeEntry("BackgroundColor", pprops->backgroundcolor);
   config->writeEntry("BackgroundPixmap", pprops->backgroundpixmap);
